Extract board representation to presenter
